Create only a single FFI, with one call to verify().
