+++ src/lxml-cffi/includes/cffi_base.py
@@ -1,0 +1,2 @@
+import cffi
+ffi = cffi.FFI()
