+++ src/lxml-cffi/includes/dtdvalid.py
@@ -1,8 +1,4 @@
-import cffi
-from . import tree
-
-ffi = cffi.FFI()
-ffi.include(tree.ffi)
+from .cffi_base import ffi
 
 ffi.cdef("""
     typedef struct _xmlValidCtxt xmlValidCtxt;
@@ -25,15 +21,3 @@ ffi.cdef("""
     xmlElementPtr	xmlGetDtdElementDesc	(xmlDtdPtr dtd, 
 					 const xmlChar * name);
 """)
-
-libxml = ffi.verify("""
-    #include "libxml/parser.h"
-    #include "libxml/valid.h"
-""",
-include_dirs=['/usr/include/libxml2'],
-libraries=['xml2'])
-
-for name in dir(libxml):
-    if name.startswith('xml'):
-        globals()[name] = getattr(libxml, name)
-
