+++ src/lxml-cffi/includes/xinclude.py
@@ -1,8 +1,5 @@
-import cffi
-from . import tree
+from .cffi_base import ffi
 
-ffi = cffi.FFI()
-ffi.include(tree.ffi)
 ffi.cdef("""
     int 	xmlXIncludeProcessTreeFlags(xmlNodePtr tree,
 					 int flags);
@@ -10,15 +7,3 @@ ffi.cdef("""
 					 int flags,
 					 void *data);
 """)
-libxml = ffi.verify("""
-    #include "libxml/xinclude.h"
-""",
-include_dirs=['/usr/include/libxml2'],
-libraries=['xml2'])
-
-def init():
-    for name in dir(libxml):
-        if name.startswith(('xml', 'XML')):
-            globals()[name] = getattr(libxml, name)
-
-init()
