+++ src/lxml/__init__.py
@@ -7,7 +7,7 @@ except ImportError:
 else:
     # PyPy prefers the cffi port
     import os
-    __path__.append(os.path.join(__path__[0], "../lxml-cffi"))
+    __path__.insert(0, os.path.join(__path__[0], "../lxml-cffi"))
 
 def get_include():
     """
